backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
cryptography==44.0.0
google-auth==2.37.0
PyJWT==2.10.1
requests==2.32.3
